Adding initial data model
