--- api/jsconfig.json
@@ -1,8 +1,0 @@
-{
-  "compilerOptions": {
-    "moduleResolution": "nodenext",
-    "target": "ES2021",
-    "checkJs": true,
-    "allowJs": true
-  }
-}
