+++ senatus/.editorconfig
