+++ senatus/config/api.ts
@@ -1,0 +1,7 @@
+export default {
+  rest: {
+    defaultLimit: 25,
+    maxLimit: 100,
+    withCount: true,
+  },
+};
