+++ senatus/src/admin/webpack.config.example.js
