+++ senatus/src/api/event/controllers/event.ts
@@ -1,0 +1,7 @@
+/**
+ * event controller
+ */
+
+import { factories } from '@strapi/strapi'
+
+export default factories.createCoreController('api::event.event');
