+++ senatus/src/api/event/routes/event.ts
@@ -1,0 +1,7 @@
+/**
+ * event router
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreRouter('api::event.event');
