+++ senatus/src/api/event/services/event.ts
@@ -1,0 +1,7 @@
+/**
+ * event service
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreService('api::event.event');
