+++ senatus/src/api/language/content-types/language/schema.json
@@ -1,0 +1,25 @@
+{
+  "kind": "collectionType",
+  "collectionName": "languages",
+  "info": {
+    "singularName": "language",
+    "pluralName": "languages",
+    "displayName": "Language",
+    "description": ""
+  },
+  "options": {
+    "draftAndPublish": true
+  },
+  "pluginOptions": {},
+  "attributes": {
+    "name": {
+      "type": "string"
+    },
+    "defaultMaxCpuTime": {
+      "type": "float"
+    },
+    "defaultMaxMemory": {
+      "type": "integer"
+    }
+  }
+}
