+++ senatus/src/api/language/controllers/language.ts
@@ -1,0 +1,7 @@
+/**
+ * language controller
+ */
+
+import { factories } from '@strapi/strapi'
+
+export default factories.createCoreController('api::language.language');
