+++ senatus/src/api/language/routes/language.ts
@@ -1,0 +1,7 @@
+/**
+ * language router
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreRouter('api::language.language');
