+++ senatus/src/api/language/services/language.ts
@@ -1,0 +1,7 @@
+/**
+ * language service
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreService('api::language.language');
