+++ senatus/src/api/problem/controllers/problem.ts
@@ -1,0 +1,7 @@
+/**
+ * problem controller
+ */
+
+import { factories } from '@strapi/strapi'
+
+export default factories.createCoreController('api::problem.problem');
