+++ senatus/src/api/problem/routes/problem.ts
@@ -1,0 +1,7 @@
+/**
+ * problem router
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreRouter('api::problem.problem');
