+++ senatus/src/api/problem/services/problem.ts
@@ -1,0 +1,7 @@
+/**
+ * problem service
+ */
+
+import { factories } from '@strapi/strapi';
+
+export default factories.createCoreService('api::problem.problem');
